requirements.txt
telethon
sqlmodel
psycopg[binary]
python-dotenv
aiohttp
